Add files via upload
Reflections:
The industrial talk has given me a lot of insider information regarding the UTMDigital services. I found the talk to be very useful as I am first year student and I got to know what kind of Network Services my university has in offer for the students which can be very pivotal for our studies.